pyproject.toml
[tool.poetry]
name = "shirokuro"
version = "0.1.0"
description = ""
authors = ["Guillermo Ferrer Bosque"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
alembic = "^1.13.1"
python = "^3.11"
httpx = "^0.27.0"
starlette = "^0.37.2"
strawberry-graphql = "^0.227.3"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.30" }
asyncpg = "^0.29.0"
psycopg = { extras = ["binary"], version = "^3.1.18" }
email-validator = "^2.1.1"
abstractcp = "^0.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-xprocess = "^0.23.0"
pytest-cov = "^4.1.0"
opencv-python = "^4.9.0.80"
pytest-asyncio = "^0.23.5.post1"
strawberry-graphql = { extras = ["debug-server"], version = "^0.227.3" }
pytest-docker-tools = "^3.1.3"
python-dotenv = "^1.0.1"
tenacity = "^8.2.3"
yapf = "^0.40.2"
rstr = "^3.2.2"
faker = "^25.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src/shirokuro"]
addopts = "--strict-markers" #--yapf --yapfdiff --pylint" #--cov --cov-append -n auto

